Added the jMonkey project to the src/ folder
You should be able to import this project with jMonkey from File->Open
Project. Hopefully that works for you guys.

If you have any challenges with the dll files, try to delete them, and see
if jMonkey will replace them with the appropriate ones.

The nbproject folder may not be necessary... but I think it has to do with
the jMonkey project being recognized by the IDE, so I'm leaving them
there.